requirements.txt
aiohttp
deep_translator
discord.py
lavalink
pickleDB
pymongo
python-dotenv
redis
